Myprofile7 머지합니당

+++ app/src/main/java/com/example/aboutme/RetrofitMyprofileData/RequestPatchProfile.kt
@@ -1,7 +1,7 @@
 package com.example.aboutme.RetrofitMyprofileData
 
 data class RequestPatchProfile(
-    val feature_id: Int,
+    val feature_id: Long,
     val feature_key: String,
     val feature_value: String
 )
